requirements.txt
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
ipyfs==0.1.3
mypy==1.2.0
mypy-extensions==1.0.0
pycodestyle==2.10.0
pydantic==1.10.7
PyYAML==6.0
requests==2.28.2
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
urllib3==1.26.15
